requirements.txt
pandas==1.3.5
python-pptx==0.6.21
shortuuid==1.0.8
google-cloud-storage==2.5.0
flask==2.2.2
Werkzeug==2.2.2